+++ apiserver/paasng/paas_wl/cnative/specs/models/app_resource.py
@@ -16,7 +16,6 @@ limitations under the License.
 We undertake not to change the open source license (MIT license) applicable
 to the current version of the project delivered to anyone in the future.
 """
-import logging
 from typing import Dict, List, Optional, Union, overload
 
 import yaml
@@ -26,18 +25,14 @@ from pydantic import ValidationError as PDValidationError
 from pydantic.error_wrappers import display_errors
 from rest_framework.exceptions import ValidationError
 
+from paas_wl.cnative.specs.constants import DEFAULT_PROCESS_NAME, ApiVersion, DeployStatus
+from paas_wl.cnative.specs.crd.bk_app import BkAppBuildConfig, BkAppProcess, BkAppResource, BkAppSpec, ObjectMetadata
 from paas_wl.platform.applications.relationship import ModuleAttrFromID, ModuleEnvAttrFromName
 from paas_wl.utils.models import BkUserField, TimestampedModel
 from paasng.engine.constants import AppEnvName
 from paasng.platform.applications.models import Application, ModuleEnvironment
 from paasng.platform.modules.constants import ModuleName
 from paasng.platform.modules.models import Module
-from paasng.utils.models import make_json_field
-
-from .constants import DEFAULT_PROCESS_NAME, ApiVersion, DeployStatus, MountEnvName, VolumeSourceType
-from .crd.bk_app import BkAppBuildConfig, BkAppProcess, BkAppResource, BkAppSpec, ObjectMetadata, VolumeSource
-
-logger = logging.getLogger(__name__)
 
 
 class AppModelResourceManager(models.Manager):
@@ -287,50 +282,3 @@ def generate_bkapp_name(obj: Union[Module, ModuleEnvironment]) -> str:
     else:
         name = f'{code}-m-{module_name}'
     return name.replace("_", "0us0")
-
-
-# TODO: 重构 models.py, 创建 models 模块, 包含 app_resource.py 和 volumes.py
-SourceConfigField = make_json_field('SourceConfigField', VolumeSource)
-
-
-class ConfigMapSourceQuerySet(models.QuerySet):
-    def get_by_mount(self, m: 'Mount'):
-        if m.source_config.configMap:
-            return self.get(module=m.module, environment_name=m.environment_name, name=m.source_config.configMap.name)
-        raise ValueError(f'Mount {m.name} is invalid: source_config.configMap is none')
-
-
-class ConfigMapSource(TimestampedModel):
-    module = models.ForeignKey('modules.Module', on_delete=models.CASCADE, db_constraint=False, null=True)
-    environment_name = models.CharField(
-        verbose_name=_('环境名称'), choices=MountEnvName.get_choices(), null=False, max_length=16
-    )
-    name = models.CharField(max_length=63)
-    data = models.TextField()
-
-    objects = ConfigMapSourceQuerySet.as_manager()
-
-    class Meta:
-        unique_together = ('module', 'name', 'environment_name')
-
-
-class Mount(TimestampedModel):
-    """挂载配置"""
-
-    module = models.ForeignKey('modules.Module', on_delete=models.CASCADE, db_constraint=False, null=True)
-    environment_name = models.CharField(
-        verbose_name=_('环境名称'), choices=MountEnvName.get_choices(), null=False, max_length=16
-    )
-    name = models.CharField(max_length=63)
-    mount_path = models.CharField(max_length=128)
-    source_type = models.CharField(choices=VolumeSourceType.get_choices(), max_length=32)
-    source_config: VolumeSource = SourceConfigField()
-
-    class Meta:
-        unique_together = ('module', 'mount_path', 'environment_name')
-
-    @property
-    def source(self) -> Union[ConfigMapSource]:
-        if self.source_type == VolumeSourceType.ConfigMap:
-            return ConfigMapSource.objects.get_by_mount(self)
-        raise ValueError(f'unsupported source type {self.source_type}')
